renamed:    src/client/chat-master/startup.ts -> client/chat-master/startup.ts 	renamed:    resources/fgamemode/client/chat-master/view/assets/index.5270b727.js -> client/chat-master/view/assets/index.5270b727.js
