+++ client/fg-external/functions.ts
