+++ client/imports.ts
@@ -1,0 +1,2 @@
+import './chat-master/startup';
+import './fg-connectplayer/connect';
