+++ fgamemode/client/chat-master/html/app.js
