+++ fgamemode/client/chat-master/startup.js
