+++ fgamemode/client/fg-external/functions.js
