+++ fgamemode/index.d.js
