+++ fgamemode/index.d.ts
