+++ fgamemode/resource.cfg
