+++ fgamemode/server/chat-master/startup.js
