+++ fgamemode/server/fg-dmarea/areas.js
@@ -1,0 +1,158 @@
+import { AreaBlip, log } from "alt-server";
+const handleAreas = [
+    {
+        x: -19.25274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 15
+    },
+    {
+        x: 224.65274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 5
+    },
+    {
+        x: 468.95274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 2
+    },
+    {
+        x: -263.16274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 44
+    },
+    {
+        x: -507.16274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 55
+    },
+    {
+        x: -751.16274658203125,
+        y: -2138.997802734375,
+        z: -60.4139404296875,
+        color: 24
+    },
+    {
+        x: -19.25274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 1
+    },
+    {
+        x: 224.65274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 22
+    },
+    {
+        x: 468.65274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 33
+    },
+    {
+        x: -263.16274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 54
+    },
+    {
+        x: -507.16274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 31
+    },
+    {
+        x: -751.16274658203125,
+        y: -1894.997802734375,
+        z: -60.4139404296875,
+        color: 55
+    },
+    {
+        x: -19.25274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 12
+    },
+    {
+        x: 224.65274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 1
+    },
+    {
+        x: 468.65274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 15
+    },
+    {
+        x: -263.16274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 22
+    },
+    {
+        x: -507.16274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 53
+    },
+    {
+        x: -751.16274658203125,
+        y: -1650.997802734375,
+        z: -60.4139404296875,
+        color: 59
+    },
+    {
+        x: -19.25274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 15
+    },
+    {
+        x: 224.65274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 26
+    },
+    {
+        x: 468.65274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 15
+    },
+    {
+        x: -263.16274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 19
+    },
+    {
+        x: -507.16274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 44
+    },
+    {
+        x: -751.16274658203125,
+        y: -1406.997802734375,
+        z: -60.4139404296875,
+        color: 55
+    }, 
+];
+new class HandleArea {
+    constructor(){
+        handleAreas.forEach((area)=>{
+            const areas = new AreaBlip(area.x, area.y, area.z, 500, 500);
+            areas.sprite = null;
+            areas.color = area.color;
+            areas.alpha = 134;
+        });
+        log('~lg~Areas was loaded');
+    }
+};
