+++ fgamemode/server/imports.js
@@ -1,0 +1,5 @@
+import './fg-connectplayer/connect';
+import './fg-commands/cmds';
+import './chat-master/startup';
+import './fg-events/death';
+import './fg-dmarea/areas';
