+++ fgamemode/shared/fg-external/functions.js
