+++ fgamemode/shared/lists/Positions.js
@@ -1,0 +1,22 @@
+export const PositionSpawnTable = [
+    {
+        x: -1368.6329345703125,
+        y: 57.125274658203125,
+        z: 53.6951904296875
+    },
+    {
+        x: -666.949462890625,
+        y: -135.25714111328125,
+        z: 37.906982421875
+    },
+    {
+        x: -676.7340698242188,
+        y: 312.79119873046875,
+        z: 83.081298828125
+    },
+    {
+        x: 52.760440826416016,
+        y: 2786.1494140625,
+        z: 57.8740234375
+    }
+];
