+++ fgamemode/shared/lists/Weapons.js
