+++ resource.cfg
@@ -1,9 +1,8 @@
 type: js,
-main: server/index.js,
-client-main: client/index.js,
+main: server/imports.js,
+client-main: client/imports.js,
 client-files: [
   "client/*"
 ],
 deps: [
-  chat
 ]
