--- resources/fgamemode/client/imports.js
@@ -1,1 +1,0 @@
-import './chat-master/startup';
