--- resources/fgamemode/resource.cfg
@@ -1,8 +1,0 @@
-type: js,
-main: server/imports.js,
-client-main: client/imports.js,
-client-files: [
-  "client/*"
-],
-deps: [
-]
