--- resources/fgamemode/server/imports.js
@@ -1,3 +1,0 @@
-import './fg-connectplayer/connect';
-import './fg-commands/cmds';
-import './chat-master/startup';
