--- server/Weapon_list.js
@@ -1,97 +1,0 @@
-export const WeaponModel = {
-	CombatPdw : 0x0A3D4D34,
-	HeavySniperMk2 : 0x0A914799,
-	HeavySniper : 0x0C472FE2,
-	SniperRifle : 0x05FC3C11,
-	FireExtinguisher : 0x060EC506,
-	CompactLauncher : 0x0781FE4A,
-	Snowball : 0x0787F0BB,
-	VintagePistol : 0x083839C4,
-	Pistol : 0x1B06D571,
-	PumpShotgun : 0x1D073A89,
-	CeramicPistol : 0x2B5EF5EC,
-	Smg : 0x2BE6766B,
-	StickyBomb : 0x2C3731D9,
-	HeavyShotgun : 0x3AABBBAA,
-	GrenadeLauncherSmoke : 0x4DD2DC56,
-	Hammer : 0x4E875F73,
-	CombatPistol : 0x5EF9FEC4,
-	MarksmanRifleMk2 : 0x6A6C02E0,
-	Railgun : 0x6D544C99,
-	BullpupRifle : 0x7F229F94,
-	Firework : 0x7F7497E5,
-	CombatMg : 0x7FD62962,
-	Flashlight : 0x8BB05FD7,
-	Mg : 0x9D07F764,
-	BullpupShotgun : 0x9D61E50F,
-	AutoShotgun : 0x12E82D3D,
-	APPistol : 0x22D8FE39,
-	Ball : 0x23C9F95C,
-	Molotov : 0x24B17070,
-	PetrolCan : 0x34A67B97,
-	Minigun : 0x42BF8A85,
-	HomingLauncher : 0x63AB0442,
-	SmgMk2 : 0x78A97CD0,
-	CarbineRifle : 0x83BF0278,
-	Crowbar : 0x84BD7BFD,
-	BullpupRifleMk2 : 0x84D6FAFD,
-	Dagger : 0x92A27487,
-	Grenade : 0x93E220BD,
-	DoubleAction : 0x97EA20B8,
-	Pistol50 : 0x99AEEB3B,
-	Knife : 0x99B507EA,
-	AssaultRifleMk2 : 0x394F415C,
-	GolfClub : 0x440E4788,
-	Raycarbine : 0x476BF155,
-	Flare : 0x497FACC3,
-	PumpShotgunMk2 : 0x555AF99A,
-	CompactRifle : 0x624FE830,
-	Nightstick : 0x678B81B1,
-	NavyRevolver : 0x917F6C8C,
-	Bat : 0x958A4A8F,
-	SpecialCarbineMk2 : 0x969C3D67,
-	StunGun : 0x3656C8C1,
-	StoneHatchet : 0x3813FC08,
-	SawnOffShotgun : 0x7846A318,
-	Wrench : 0x19044EE0,
-	MicroSmg : 0x13532244,
-	FlareGun : 0x47757124,
-	Gusenberg : 0x61012683,
-	SNSPistolMk2 : 0x88374054,
-	PoolCue : 0x94117305,
-	BzGas : 0xA0973D5E,
-	Musket : 0xA89CB99E,
-	GrenadeLauncher : 0xA284510B,
-	Unarmed : 0xA2719263,
-	ProxMine : 0xAB564B93,
-	AdvancedRifle : 0xAF113F99,
-	Raypistol : 0xAF3696A1,
-	Rpg : 0xB1CA77B1,
-	Rayminigun : 0xB62D1F67,
-	PipeBomb : 0xBA45E8B8,
-	HazardCan : 0xBA536372,
-	MiniSmg : 0xBD248B55,
-	SNSPistol : 0xBFD21232,
-	PistolMk2 : 0xBFE256D4,
-	AssaultRifle : 0xBFEFFF6D,
-	SpecialCarbine : 0xC0A3098D,
-	Revolver : 0xC1B3C3D1,
-	MarksmanRifle : 0xC734385A,
-	RevolverMk2 : 0xCB96392F,
-	BattleAxe : 0xCD274149,
-	Knuckle : 0xD8DF3C3C,
-	HeavyPistol : 0xD205520E,
-	MachinePistol : 0xDB1AA450,
-	CombatMgMk2 : 0xDBBD7280,
-	MarksmanPistol : 0xDC4DB296,
-	Machete : 0xDD5DF8D9,
-	Switchblade : 0xDFE37640,
-	AssaultShotgun : 0xE284C527,
-	DbShotgun : 0xEF951FBB,
-	AssaultSmg : 0xEFE7E2DF,
-	Hatchet : 0xF9DCBF2D,
-	Bottle : 0xF9E6AA4B,
-	CarbineRifleMk2 : 0xFAD1F1C9,
-	Parachute : 0xFBAB5776,
-	SmokeGrenade : 0xFDBC8A50
-};
