+++ shared/fg-external/functions.ts
