+++ shared/lists/Weapons.ts
