--- src/client/imports.ts
@@ -1,1 +1,0 @@
-import './chat-master/startup';
